feat: Introduce macros for skipping tests
- add `catalog show XXX` command to check catalog configuration for the
  service
- add `skip_without_service` and `skip_without_extension` macros to
  allow tests requiring certain cloud features. Since there is no way to
  simply skip test in rust return OK.
