+++ openstack_cli/tests/block_storage/v3/attachment/list.rs
@@ -11,17 +11,18 @@
 // limitations under the License.
 //
 // SPDX-License-Identifier: Apache-2.0
+//
+// WARNING: This file is automatically generated from OpenAPI schema using
+// `openstack-codegenerator`.
 
 use assert_cmd::prelude::*;
 use std::process::Command;
 
 #[test]
 fn list() -> Result<(), Box<dyn std::error::Error>> {
-    skip_without_service!("object-store");
-
     let mut cmd = Command::cargo_bin("osc")?;
 
-    cmd.args(["object-store", "container", "list"]);
+    cmd.args(["block-storage", "attachment", "list"]);
     cmd.assert().success();
 
     Ok(())
