+++ openstack_cli/tests/macros.rs
@@ -15,17 +15,28 @@
 // WARNING: This file is automatically generated from OpenAPI schema using
 // `openstack-codegenerator`.
 
-use assert_cmd::prelude::*;
-use std::process::Command;
+#[macro_export]
+macro_rules! skip_without_extension {
+    ($srv:literal, $ext:literal) => {
+        if Command::cargo_bin("osc")?
+            .args([$srv, "extension", "show", $ext])
+            .ok()
+            .is_err()
+        {
+            return Ok(());
+        }
+    };
+}
 
-#[test]
-fn list() -> Result<(), Box<dyn std::error::Error>> {
-    skip_without_extension!("network", "security-groups-default-rules");
-
-    let mut cmd = Command::cargo_bin("osc")?;
-
-    cmd.args(["network", "default-security-group-rule", "list"]);
-    cmd.assert().success();
-
-    Ok(())
+#[macro_export]
+macro_rules! skip_without_service {
+    ($srv:literal) => {
+        if Command::cargo_bin("osc")?
+            .args(["catalog", "show", $srv])
+            .ok()
+            .is_err()
+        {
+            return Ok(());
+        }
+    };
 }
